recommendationsystem-backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic
python-dotenv
openai
pyautogen
